optimize code style
